@@ -9,60 +9,60 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "postinstall": "ngcc --tsconfig ./tsconfig.app.json"
+    "postinstall": ""
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/service-worker": "^17.3.12",
+    "@angular/animations": "^18.2.7",
+    "@angular/cdk": "^18.2.7",
+    "@angular/common": "^18.2.7",
+    "@angular/compiler": "^18.2.7",
+    "@angular/core": "^18.2.7",
+    "@angular/forms": "^18.2.7",
+    "@angular/platform-browser": "^18.2.7",
+    "@angular/platform-browser-dynamic": "^18.2.7",
+    "@angular/router": "^18.2.7",
+    "@angular/service-worker": "^18.2.7",
     "@auth0/angular-jwt": "^5.1.2",
-    "@fortawesome/angular-fontawesome": "^0.10.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@types/crypto-js": "^3.1.47",
-    "chart.js": "^2.9.4",
-    "core-js": "^2.6.12",
-    "crypto-js": "^3.3.0",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@types/crypto-js": "^4.2.2",
+    "chart.js": "^4.4.4",
+    "core-js": "^3.38.1",
+    "crypto-js": "^4.2.0",
     "font-awesome": "^4.7.0",
     "howler": "^2.2.3",
     "moment": "^2.29.4",
     "ng2-currency-mask": "^13.0.3",
     "primeflex": "^3.3.1",
-    "primeicons": "^6.0.1",
+    "primeicons": "^7.0.0",
     "primeng": "^17.18.11",
-    "quill": "^1.3.7",
-    "rxjs": "^6.6.7",
+    "quill": "^2.0.2",
+    "rxjs": "^7.8.1",
     "toasty": "^6.0.3",
-    "tslib": "^1.10.0",
-    "zone.js": "^0.14.10"
+    "tslib": "^2.7.0",
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
-    "@angular/cli": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.7",
+    "@angular/cli": "^18.2.7",
+    "@angular/compiler-cli": "^18.2.7",
+    "@angular/language-service": "^18.2.7",
     "@types/howler": "^2.2.7",
-    "@types/jasmine": "~2.5.53",
+    "@types/jasmine": "^5.1.4",
     "@types/jasminewd2": "^2.0.10",
-    "@types/node": "^15.14.9",
-    "jasmine-core": "3.8",
-    "jasmine-spec-reporter": "~5.0.0",
+    "@types/node": "^22.7.5",
+    "jasmine-core": "^5.3.0",
+    "jasmine-spec-reporter": "^7.0.0",
     "karma": "~6.4.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-cli": "~1.0.1",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ts-node": "^3.2.0",
-    "tslint": "~6.1.0",
-    "typescript": "5.4.5"
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "ts-node": "^10.9.2",
+    "tslint": "^5.20.1",
+    "typescript": "^5.5.0"
   }
 }
